URL Encode Search / line breaks
Wrapped lines in the search button to ~80 char
URLEncode the search string
Changed the trailing " " to "+" in the GitHub search
Added code to directly perform the LC Dictionary search (will not require the URL method)
Added local for the Wiki search string

Stack behavior doesn't seem to execute when running from the GitHub repo.
The `emptyUrls` handler isn't firing.  Made adjustments to the code that
are needed for the other changes.